// src/features/lidar/viewer/renderer/types.ts
export interface HeightmapParams {
  data: Float32Array;
  width: number;
  height: number;
  originX: number;
  originZ: number;
  scaleX: number;
  scaleZ: number;
}

export interface SnowParams {
  data: Float32Array;
  width: number;
  height: number;
  originX: number;
  originZ: number;
  scaleX: number;
  scaleZ: number;
}